pyproject.toml
[project]
name = "fasthtml-components"
version = "0.1.0"
description = "FastHTML + MonsterUI application template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.12.0",
    "monsterui>=1.0.35",
]

[tool.ruff]
# Ignore undefined name errors from star imports
ignore = ["F405", "F403"]